ovn-controller.at: Fix test for debug/delay-nb-cfg-report.
In the test case for debug/delay-nb-cfg-report, it sets the delay as
2 seconds, and firstly triggered a sync with wait timeout being 1 sec
and expects the wait timedout, and then it immediately started another
sync with wait timeout being 3 sec, expecting it not timing out.

However, since 2 + 2 = 1 + 3, there is a good chance that the second
wait would still time out, so the test fails occasionally. This patch
fixes it by sleep another 1 sec to let the 1st delay complete before
starting the second sync and wait.

Reported-by: Ilya Maximets <i.maximets@ovn.org>
Fixes: a8182e21484 ("ovn-controller: Unix command debug/delay-nb-cfg-report.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>