ovn-controller: Handle changes in requested-tnl-key
runtime_data_sb_datapath_binding_handler() only handles deletion of rows
in the Datapath_Binding table in OVN-SB. The user can update the
requested-tnl-key by the following command:

ovn-nbctl set logical_switch sw0 other_config:requested-tnl-key=<key>

This command modifies the tunnel_key column. This patch
ensures that an update of this column is handled by the
incremental processing engine by forcing a recompute
of the runtime_data node.

As it is expected that changing the requested-tnl-key is not updated
frequently, we do not attempt to make this update incrementally.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>