ovn-northd: Fix router policy pkt mark over flow if the value is greater than signed int.
If the value of pkt_mark in the router policy options is greater than 2147483647, we
are ignoring it. This is because we use smap_get_int(). This patch fixes this issue
by adding a wrapper function - ovn_smap_get_uint(). The OpenvSwitch library doesn't have
this smap helper function.

ovn_smap_get_uint() needs to be removed once the OpenvSwitch library has this helper
function. A comment has been added in the code for this.

Fixes: a123ef0fb8fd("Support packet metadata marking for logical router policies.")
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=1878248
Reported-by: Alexander Constantinescu <aconstan@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>