ovn-northd: Add localnet ports to Multicast_Groups created by IGMP_Group.
In case a logical switch is a "bridged logical switch", i.e., it has a
localnet port, add the localnet port to all multicast groups generated
from IGMP_Groups.  This is needed because multicast packets are never
tunneled between hypervisors.

Note: There still exists an issue regarding IP Multicast Relay (routing)
when traffic is received on bridged logical switches connected to
logical routers with mcast_relay=true.  That issue is not addressed
yet but an item is added to TODO.rst to track it.

Reported-at: https://bugzilla.redhat.com/1886103
Fixes: 5d1527b11e94 ("ovn-northd: Add IGMP Relay support")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>