actions: Fix issues with fwd_group action.
The example for the fwd_group action was obviously wrong (it needed a
closing parenthesis and semicolon) and then when I looked closer I saw
other issues:

  - It's weird to insist on "true" or "false" in quotation marks (why?)
    and the example didn't show that.

  - The example didn't show that port names need to be quoted.

  - The implementation didn't allow specifying liveness=false (why?)
    or liveness="false" and the test even checked for that.

This fixes all that stuff.  For backward compatibility, it still accepts
"true" (and "false") in quotation marks, but now accepts without as well
for future use.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>