ovn-controller: Propagate nb_cfg to the local OVS DB.
The NB.NB_Global.nb_cfg value gets propagated to
Chassis_Private.nb_cfg (and then to NB.NB_Global.hv_cfg) as soon as
ovn-controller has finished installing OVS flows corresponding to
the NB DB state.

However, if the CMS runs monitoring applications on the chassis itself,
in order to detect that the NB changes have been applied, it has to
connect to the NB/SB database.  In a scaled deployment this additional
connection might induce performance issues.

In order to avoid that we now (also) propagate nb_cfg to the local OVS
DB, in the record corresponding to the integration bridge in table
Bridge, as external_id "ovn-nb-cfg".

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>