controller: Allow pinctrl thread to handle packet-ins when version mismatch with northd.
The commit 1dd27ea7aea4 added the support to pin ovn-controller to a
specific version of ovn-northd.  This patch did not handle the
scenario the packet-in scenario when ovn-controller is started and it
detects version mismatch with ovn-northd.  In this case, pinctrl
thread is not configured with the proper integration bridge name,
because of which it cannot handle any packet-ins.

This patch addresses this scenario.  This is required so that any
existing VMs do not loose DHCP address during renewal.

Fixes: 1dd27ea7aea4("Provide the option to pin ovn-controller and ovn-northd to a specific version")
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>