vtep tests: Fix the failing test - ovn-controller-vtep - binding 1.
The below check in the test - "ovn-controller-vtep - binding 1." is
failing after the commit [1].

Before the commit[1], the check was -
  ovn-sbctl --timeout=10 wait-until Port_Binding br-vtep_lswitch0_dup chassis!='[[]]'

And after this commit, the check is -
  check ovn-sbctl wait-until Port_Binding br-vtep_lswitch0_dup chassis!='[[]]'

The test expects 'br-vtep_lswitch0_dup' to be claimed by the chassis,
but there is already a logical port - 'br-vtep_lswitch0' which is
claimed by the same chassis. Both these logical ports are set with the
same options, which means the lport - 'br-vtep_lswitch0_dup' is a
duplicate and it should not be claimed. The test also checks if there is
warning message from ovn-controller-vtep about it.

So looks like the test case was wrong from the beginning itself.  The
"ovn-sbctl --timeout=10 .. " command before the commit [1] was also failing,
but there was no check to fail the test before. But the commit [1] makes the
test fail because its wrapped with 'check' helper function.

This patch removes this wrong test check.

[1] - 5fedcce6c3e0("tests: Remove explicit timeouts from most utility invocations.")
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>