ovn-controller: Relax get_nb_cfg() condition seqno check.
The IDL might decide to resend pending monitor condition changes
implicitly (one such case is at reconnect) in which case the client
(ovn-controller) might end waiting for a condition seqno that has
already been satisfied.

In order to avoid this case, we now relax the seqno check in
get_nb_cfg() and if the IDL condition seqno is greater or equal than
what ovn-controller expected after changing monitor conditions it means
that there are no in-flight pending changes.

Fixes: 58e5d32b011b ("ovn-controller: Fix nb_cfg update with monitor_cond_change in flight.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>