nbctl: Use partial set updates instead of re-setting the whole column.
Northbound database has many tables that could contain columns with
very large sets.  For example 'ports' column of the 'Logical_Switch'
table could contain thousands of ports in a set.

Current strategy of nbctl while adding a new port to the set is to
copy all existing ports, add one and set the new set of ports.
Similar behavior is for deletion too.

If we have 1000 ports and want to add one more, resulted transaction
in current code will look like this:

  {transact,
        < wait operation >
        < create new lsp in Logical_Switch_Port table >

        where: _uuid  == 'logical switch row uuid'
        op   : update
        rows : ports ['< list of current 1000 ports + 1 new >']
  }

The code was written before support for partial updates for sets was
implemented.  Now we have it and can replace the old strategy.
With this change resulted transaction will be:

  {transact,
        < wait operation >
        < create new lsp in Logical_Switch_Port table >

        where: _uuid  == 'logical switch row uuid'
        op   :  mutate
        mutations : ports insert ['< 1 uuid of a new lsp >']
  }

Unfortunately, for now, this only decreases transaction size in half,
because '<wait operation>', that is still in the transaction, contains
the full current list of ports:

        where: _uuid  == 'logical switch row uuid'
        op   : wait
        until: ==
        rows : ports ['< list of current 1000 ports >']

But anyway, beside the overall code cleanup, this reduces transaction
size in half and should reduce pressure on the ovsdb-server since it
will not need to parse and process all 1000 ports twice.

This change doesn't affect 'append_request' messages within the raft
cluster, because ovsdb-server is not smart enough to use mutations
there, and this will also not affect 'update' messages from the
ovsdb-server to its clients, because it is smart enough to construct
'modify' updates regardless of the original transaction.

One difference between full updates and partial is that partial
changes are not visible for the idl client until transaction is
committed and the update received from the server.  New switch ports
are not visible while iterating over 'ports' of the logical switch and
removed ports are still there.  For this reason, we have to maintain
runtime cache of the mapping between ports and routers/switches they
attached to.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>