northd: Fix datapath swapping in logical flows.
Commit that introduced support for logical datapath groups removed
the 'logical_datapath' column from the lflow hash.  If datapath groups
disabled, there will be many flows that are same except for the
'logical_datapath' column, so they will have exactly same hash.
Since 'logical_datapath' is not involved into comparison inside
ovn_lflow_equal(), all these flows will be considered equal.

While iterating over Southbound DB records, ovn_lflow_found() will
return first of many "equal" flows and, likely, not the right one.
Comparison of logical datapaths will say that we need to update
the logical datapath, so it will be updated with the value from
the flow that we just found.  Flow that we found will be removed from
the hash map.  Similar procedure for the next DB record will give
us different "equal" flow leading to the update of the
'logical_datapath' column for the next record.  And so on.  This
way we will swap 'logical_datapath' column for almost all logical
flows.  Since we're not using same lflow more than once, resulted
database will still be correct, but all these unnecessary steps will
generate huge database transaction if we'll have at least one new
or modified logical flow, because that will cause different order
in which lflows will be found in a hash map.

Fix that by re-adding 'logical_datapath' column back to hash and to
the check for equality.  This way correct lflows could be found, so
there will be no unnecessary updates.

Some functions and variables renamed to better describe their meaning.
Also size_t replaced with uint32_t to avoid confusion.  lib/hash.c
always returns uint32_t as a hash value and that is important because
we will want to update current hash with the datapath value and not to
re-calculate it for all lflow columns.

Fixes: bfed22400675 ("northd: Add support for Logical Datapath Groups.")
Reported-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>