pinctrl: Fix race condition when explicitly clearing IGMP groups.
When a user flushes the IGMP Groups ("ovn-sbctl ip-multicast-flush")
there's no guarantee that the ovn-controller main thread runs soon after
the pinctrl thread has flushed the in-memory IGMP groups, in
ip_mcast_snoop_flush().

To avoid unnecessarily waking the main thread just clear the local
IGMP_Group records in the main thread.

This scenario is quite hard to encounter in real deployments but every
now and then we hit it in CI.

Fixes: 70ff8243040f ("OVN: Add IGMP SB definitions and ovn-controller support")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>