controller: Implement a generic barrier based on ofctrl cur_cfg sync.
A new module, 'ofctrl-seqno', is added to implement this generic
barrier.  Other modules can register their own types of seqno update
requests.  The barrier implementation ensures that the a seqno update
request is acked (returned by ofctrl_acked_seqnos_get()) only if the
OVS flow operations that have been requested when the seqno update
request was queued have been processed by OVS.

For now, the only user of this barrier is the main ovn-controller
module but a future commit will use it too in order to mark
Port_Bindings and OVS interfaces as "fully installed".

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>