ovn-nbctl: add --bfd option to lr-route-add
Introduce the --bfd option to lr-route-add command.
If the BFD session UUID is provided, it will be used for the OVN route
otherwise the next-hop will be used to perform a lookup in the OVN BFD
table.
If the lookup fails and outport is specified, a new entry in the BFD table
will be created using the nexthop as dst_ip and outport as logical_port.

Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>