lflow: Refactor convert_match_to_expr() to explicitly consume prereqs.
It was (and still is) the responsibility of the caller of
convert_match_to_expr() to explicitly free any 'prereqs' expression that
was passed as argument if the expression parsing of the 'lflow' match failed.

However, convert_match_to_expr() now updates the value of '*prereqs' setting
it to NULL in the successful case.  This makes it easier for callers of the
function because 'expr_destroy(prereqs)' can now be called unconditionally.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>