ovn-northd: Simplify iteration through ACLs.
The existing code here seemed like an expensive way to find every ACL.
This simplifies it.

There is one behavioral change here: if an ACL is only referred to by
a Port_Group that itself has no references, then the old code will not
process it and the new code will.  I don't know whether that was why
the code was written the way it was.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>